requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-environ==0.11.2
requests==2.31.0
python-telegram-bot==20.8